% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamapr.R
\docType{package}
\name{deltamapr}
\alias{deltamapr-package}
\alias{deltamapr}
\title{deltamapr: A data package for spatial data relevant to the upper San Francisco Estuary}
\description{
This package contains a number of spatial datasets in sf format
}
\section{Prefixes}{

\emph{Prefixes to data object names are used to denote the data type.}
\describe{
\item{R}{Regions. Sets of regions, strata, or subregions.}
\item{WW}{Waterways. These are shapefiles of water coverage.}
\item{H}{Habitats. These are shapefiles with information on habitat types or environment.}
\item{P}{Points. These are shapefiles with information on long-term monitoring sampling locations.}
}
}

\author{
\strong{Maintainer}: Samuel M Bashevkin \email{sam.bashevkin@waterboards.ca.gov} (\href{https://orcid.org/0000-0001-7406-7089}{ORCID})

Authors:
\itemize{
  \item Arthur Barros \email{Arthur.Barros@Wildlife.ca.gov}
  \item Rosemary Hartman \email{Rosemary.Hartman@water.ca.gov}
}

}
