% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H_CARI_streams}
\alias{H_CARI_streams}
\title{California Aquatic Resources Inventory streams}
\format{
a sf tibble with 55,522 rows and 16 columns.
\describe{
\item{gid}{a placeholder attribute for EcoAtlas.org (unused).}
\item{name}{The name of a particular wetland feature. (e.g. Pajaro River).}
\item{orig_class}{Original classification of the wetland in the source dataset. Source datasets (e.g. NWI, BAARI) use different classification systems. The 'orig_class' field preserves that information.}
\item{source_dat}{a description of the imagery or other primary data source from which the wetland feature was identified and mapped.}
\item{clickcode}{An alphanumeric code for the wetland classification. Clickcode values are undergoing revision and will be updated in a future release of CARI.}
\item{clicklabel}{a detailed description of a feature's wetland type. Clicklabel is the most detailed classification provided by CARI.
Original classifications from component datasets (see 'orig_class'' above) are translated or "crosswalked" into CARI's classification
system. Clicklabel values include: "Fluvial Natural", "Fluvial Unnatural", "Tidal Natural", and "Tidal Unnatural".}
\item{legheader}{In EcoAtlas.org, 'legheader' is the value that appears in the legend heading. "Drainage Features" is the only value for legheader.}.
\item{legcode}{a 1 or 2 letter code identifying whether the major stream class has tidal influence (TR) or not (R), associated with
the 'leglabellevel2' field. Provides less detail than 'clickcode'.}
\item{leglabel}{Common terminology for a feature's wetland type. In EcoAtlas.org, Leglabel is the value that appears in the legend,
when you click on a particular feature to identify it, and in analyses of wetland area generated by EcoAtlas' Landscape Profile tool.
Leglabel values include:"Fluvial" and "Tidal".}
\item{dataset_org}{the agency or organization that originally mapped the wetland feature.}
\item{orig_datas}{the original dataset (and data publication date if known) that was integrated into CARI.}
\item{ReleaseDate}{The date that a particular feature was integrated into the CARI dataset, including updated versions of a particular feature.}
\item{lastupdate}{The date that a particular feature was updated.}
\item{CARI_id}{Unique id for that CARI feature.}
\item{Shape_Length}{Length of feature in internal units.}
\item{Shape}{Geometry column.}

}
}
\usage{
H_CARI_streams
}
\description{
Stream data from the California Aquatic Resources Inventory (CARI), cut to only retain feature that touch a 5km buffer around the Bay-Delta (as defined by the \code{\link{WW_Delta}} dataset).
}
\section{Credit}{

This dataset was created by the San Francisco Estuary Institute and is available (along with full metadata) \href{https://www.sfei.org/data/california-aquatic-resource-inventory-cari-version-03-gis-data}{here}.
}

\keyword{datasets}
