% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H_CARI_wetlands}
\alias{H_CARI_wetlands}
\title{California Aquatic Resources Inventory wetlands}
\format{
a sf tibble with 37,135 rows and 18 columns.
\describe{
\item{orig_class}{Original classification of the wetland in the source dataset. Source datasets (e.g. NWI, BAARI) use different classification systems. The 'orig_class' field preserves that information.}
\item{name}{The name of a particular wetland feature. (e.g. Alpine Lake).}
\item{globalid}{A placeholder attribute for EcoAtlas.org (unused).}
\item{source_data}{A description of the imagery or other primary data source from which the wetland feature was identified and mapped.}
\item{organization}{The agency or organization that originally mapped the wetland feature.}
\item{orig_dataset}{The original dataset (and data publication date if known) that was integrated into CARI.}
\item{lastupdate}{The date that a particular feature was updated.}
\item{clickcode}{An alphanumeric code for the wetland classification. Clickcode values are undergoing revision and will be updated in a future release of CARI.}
\item{clicklabel}{A detailed description of a feature's wetland type. Clicklabel is the most detailed classification provided by CARI.
Original classifications from component datasets (see 'orig_class'' above) are translated or "crosswalked" into CARI's classification
system. While CARI's component datasets may provide more detail for certain types of wetlands, CARI provides a single
statewide classification for wetland types. E.g. "Depressional Perennial Natural Emergent". Major classes within the 'clicklabel'
field include: wetlands and (deep) open water. Classes and types include: depessions, playas, estuarine, lacustrine. marine,
riverine, lagoon, and slope.}
\item{legcode}{a 1-3 letter code signifying the major wetland class, associated with the 'leglabellevel2' field. Provides less detail than 'clickcode'.}
\item{leglabellevel1}{Common terminology for a feature's wetland type. In EcoAtlas.org, Leglabellevel1 is the value that appears in
the legend, when you click on a particular feature to identify it, and in analyses of wetland area generated by EcoAtlas' Landscape
Profile tool.}
\item{leglabellvel2}{Major classification (less detailed) of the wetland classification provided by the 'clicklabel' field. E.g. "Depressional".}
\item{legend_headings}{In EcoAtlas.org, 'legend_headings' is the value that appears in the legend heading.}
\item{ReleaseDate}{the date that a particular feature was integrated into the CARI dataset, including updated versions of a particular feature.}
\item{CARI_id}{Unique id for that CARI feature.}
\item{Shape_Length}{Length of feature in internal units.}
\item{Shape_Area}{Area of feature in internal units squared.}
\item{Shape}{Geometry column.}
}
}
\usage{
H_CARI_wetlands
}
\description{
Wetlands habitat data from the California Aquatic Resources Inventory (CARI), cut to only retain feature that touch a 5km buffer around the Bay-Delta (as defined by the \code{\link{WW_Delta}} dataset).
}
\section{Credit}{

This dataset was created by the San Francisco Estuary Institute and is available (along with full metadata) \href{https://www.sfei.org/data/california-aquatic-resource-inventory-cari-version-03-gis-data}{here}.
}

\keyword{datasets}
