% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P_RKI}
\alias{P_RKI}
\title{P_RKI}
\format{
a sf tibble with 2095 rows and 5 columns.
\describe{
\item{NAME}{Name of Water Body.}
\item{WaterBodyType}{Type of water body, either River, Canal, Slough, Cut, Island, Bay, Channel, Creek, Lake, Shore, Aqueduct, or Cuttoff.}
\item{ID}{ID given to each point, a concatenation of the abbreviated water body name and river kilometer.}
\item{RKI}{River Kilometer, as measured from mouth of the water body.}
\item{geometry}{Shapefile point coordinates.}
}
}
\usage{
P_RKI
}
\description{
Points showing official river kilometers as used by the CA Department of Water Resources for major bays, sloughs, and rivers in the San Francisco Estuary.
}
\keyword{datasets}
