% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_Delta}
\alias{R_Delta}
\title{Legal Delta Boundary}
\format{
a sf tibble with 1 row and 9 columns.
\describe{
\item{OBJECTID}{OBJECTID}
\item{comments}{comments}
\item{Source}{Source}
\item{Date_Recor}{Date_Recor}
\item{Edited_By}{Edited_By}
\item{Date_Data_}{Date_Data_}
\item{GlobalID}{GlobalID}
\item{Shape__Are}{Shape__Are}
\item{Shape__Len}{Shape__Len}
\item{geometry}{Shapefile polygon coordinates.}
}
}
\usage{
R_Delta
}
\description{
Boundary of the Legal Delta
}
\section{Credit}{

This shapefile was published by the California Department of Water Resources and can be found along with metadata in the \href{https://data.cnra.ca.gov/dataset/i03-legaldeltaboundary1}{data publication}.
}

\keyword{datasets}
