% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_EDSM_Strata}
\alias{R_EDSM_Strata}
\alias{R_EDSM_Strata_1617P1}
\alias{R_EDSM_Strata_17P2}
\alias{R_EDSM_Strata_17P3}
\alias{R_EDSM_Strata_1718P1}
\alias{R_EDSM_Strata_18P23}
\alias{R_EDSM_Strata_1819P1}
\alias{R_EDSM_Strata_19P3}
\title{EDSM Strata}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 4 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 5 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 6 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
R_EDSM_Strata_1617P1

R_EDSM_Strata_17P2

R_EDSM_Strata_17P3

R_EDSM_Strata_1718P1

R_EDSM_Strata_18P23

R_EDSM_Strata_1819P1

R_EDSM_Strata_19P3
}
\description{
Strata polygons from the Enhanced Delta Smelt Monitoring Program
}
\section{Naming}{

Files are named with the year(s) then phase(s). So \code{R_EDSM_Strata_1617P1} refers to the 2016-2017 Phase 1 strata,
while \code{R_EDSM_Strata_18P23} refers to the 2018 Phase 2 and Phase 3 strata.
}

\section{Duplication}{

Sometimes the same strata were used in multiple years. In this case, only the first year that set of strata appeared is made available and all years and phases that dataset
applies to are listed below.
\describe{
\item{R_EDSM_Strata_18P23}{Equivalent to 2019 Phase 2 and 2020 Phase 2-3 Strata}
\item{R_EDSM_Strata_1819P1}{Equivalent to 2019-2020 Phase 1 Strata}
}
}

\section{Credit}{

These datasets were created by the United States Fish and Wildlife Service for the
\href{https://www.fws.gov/lodi/juvenile_fish_monitoring_program/jfmp_index.htm}{Enhanced Delta Smelt Monitoring Program}.
}

\seealso{
\code{\link{R_EDSM_Regions}}, \code{\link{R_EDSM_Subregions}}
}
\keyword{datasets}
