% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_EDSM_Subregions}
\alias{R_EDSM_Subregions}
\alias{R_EDSM_Subregions_1617P1}
\alias{R_EDSM_Subregions_17P2}
\alias{R_EDSM_Subregions_1718P1}
\alias{R_EDSM_Subregions_18P23}
\alias{R_EDSM_Subregions_1819P1}
\alias{R_EDSM_Subregions_19P3}
\title{EDSM Subregions}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 29 rows and 11 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 29 rows and 11 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 32 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 14 rows and 6 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 26 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 18 rows and 4 columns.
}
\usage{
R_EDSM_Subregions_1617P1

R_EDSM_Subregions_17P2

R_EDSM_Subregions_1718P1

R_EDSM_Subregions_18P23

R_EDSM_Subregions_1819P1

R_EDSM_Subregions_19P3
}
\description{
Strata polygons from the Enhanced Delta Smelt Monitoring Program
}
\section{Naming}{

Files are named with the year(s) then phase(s). So \code{R_EDSM_Subregions_1617P1} refers to the 2016-2017 Phase 1 subregions,
while \code{R_EDSM_Subregions_18P23} refers to the 2018 Phase 2 and Phase 3 subregions.
}

\section{Duplication}{

Sometimes the same subregions were used in multiple years. In this case, only the first year that set of subregions appeared is made available and all years and phases that dataset
applies to are listed below.
\describe{
\item{R_EDSM_Subregions_1617P1}{Equivalent to 2017 Phase 3 Subregions}
\item{R_EDSM_Subregions_18P23}{Equivalent to 2019 Phase 2 and 2020 Phase 2-3 Subregions}
\item{R_EDSM_Subregions_1819P1}{Equivalent to 2019-2020 Phase 1 Subregions}
}
}

\section{Credit}{

These datasets were created by the United States Fish and Wildlife Service for the
\href{https://www.fws.gov/lodi/juvenile_fish_monitoring_program/jfmp_index.htm}{Enhanced Delta Smelt Monitoring Program}.
}

\seealso{
\code{\link{R_EDSM_Regions}}, \code{\link{R_EDSM_Strata}}, \code{\link{R_EDSM_Subregions_Mahardja}}, \code{\link{R_EDSM_Subregions_Mahardja_FLOAT}}
}
\keyword{datasets}
