% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_EDSM_Subregions_Mahardja_FLOAT}
\alias{R_EDSM_Subregions_Mahardja_FLOAT}
\title{Bay-Delta regions}
\format{
a sf tibble with 40 rows and 4 columns.
\describe{
\item{Region}{Region}
\item{SubRegion}{SubRegion.}
\item{SQM}{Area of subregion in square miles.}
\item{geometry}{Shapefile polygon coordinates.}
}
}
\usage{
R_EDSM_Subregions_Mahardja_FLOAT
}
\description{
A modified version of \code{\link{R_EDSM_Subregions_Mahardja}} with the "Georgiana Slough" and "Sacramento River near Ryde" SubRegions merged, retaining the "Sacramento River near Ryde" name.
This version was built for the IEP FLOAT/Drought MAST analyses in 2021
}
\section{Credit}{

This shapefile was created by the United States Fish and Wildlife Service for the Enhanced Delta Smelt Monitoring Program, modified by Brian Mahardja, then modified by Sam Bashevkin.
}

\seealso{
\code{\link{R_EDSM_Subregions}}, \code{\link{R_EDSM_Regions}}, \code{\link{R_EDSM_Strata}}, \code{\link{R_EDSM_Subregions_Mahardja}}
}
\keyword{datasets}
