% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_Suisun}
\alias{R_Suisun}
\title{Suisun Marsh Boundary}
\format{
a sf tibble with 1 row and 9 columns.
\describe{
\item{OBJECTID}{OBJECTID}
\item{LOCATION}{LOCATION}
\item{GlobalID}{GlobalID}
\item{Shape__Are}{Shape__Are}
\item{Shape__Len}{Shape__Len}
\item{geometry}{Shapefile polygon coordinates.}
}
}
\usage{
R_Suisun
}
\description{
Boundary of Suisun Marsh
}
\section{Credit}{

This shapefile was published by the California Department of Water Resources and can be found along with metadata in the \href{https://data.cnra.ca.gov/dataset/i03-suisunmarshboundary1}{data publication}.
}

\keyword{datasets}
