% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stations}
\alias{Stations}
\title{Stations}
\format{
a sf tibble with 954 rows and 7 columns.
\describe{
\item{Station}{Code used for station.}
\item{StationName}{Full name of station (if applicable).}
\item{Latitude}{Latitude in decimal degrees, WGS84.}
\item{Longitude}{Longitude in decimal degrees, WGS84.}
\item{Parameter}{Type of data collected.}
\item{Source}{Name of sampling program.}
\item{geometry}{Shapefile point coordinates.}
}
}
\usage{
Stations
}
\description{
Shapefile with regular sampling locations for most IEP surveys used between 2015 and 2021
}
\keyword{datasets}
