% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WW_Canals}
\alias{WW_Canals}
\title{Canals}
\format{
a sf tibble with 509 rows and 7 columns.
\describe{
\item{Name}{Name of canal or aquaduct}
\item{Operator}{Agency responsible for canal or aquaduct}
\item{Conv_Type}{Type of conveyance. Either 'pipeline', 'tunnel', 'canal', or 'pool'. Only available for SWP and Reclamation datasets.}
\item{Conv_Sub}{Name of subdivision of canal (SWP only)}
\item{Comments}{Comments or additional information provided about the data}
\item{Shape_Leng}{Shape Length}
\item{geometry}{Shapefile polygon coordinates.}
}
}
\usage{
WW_Canals
}
\description{
Shapefile of canals run by the State Water Project, Bureau of Reclamation, and local water agencies.
}
\section{Credit}{

This shapefile was produced by the California Department of Water Resources.
}

\keyword{datasets}
