% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WW_DBW}
\alias{WW_DBW}
\title{DBW waterways}
\format{
a sf tibble with 321 rows and 16 columns.
}
\usage{
WW_DBW
}
\description{
Shapefile of waterways in the San Francisco Estuary extending into the Southern Central Valley from the California State Parks Division of Boating and Waterways.
}
\section{Credit}{

This shapefile was produced by the California State Parks Division of Boating and Waterways.
}

\keyword{datasets}
