% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WW_Delta}
\alias{WW_Delta}
\title{Delta waterways}
\format{
a sf tibble with 282 rows and 10 columns.
\describe{
\item{AREA}{Area.}
\item{PERIMETER}{Perimeter.}
\item{HYDRO_POLY}{HYDRO_POLY.}
\item{HYDRO_PO_1}{HYDRO_PO_1.}
\item{HYDRO_24K_}{HYDRO_24K_.}
\item{HNAME}{HNAME.}
\item{Shape_Leng}{Shape_Length}
\item{Shape_Area}{Shape_Area.}
\item{geometry}{Shapefile polygon coordinates.}
}
}
\usage{
WW_Delta
}
\description{
Shapefile of delta waterways
}
\section{Credit}{

This shapefile was created by the California Department of Fish and Wildlife
}

\keyword{datasets}
