% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WW_Watershed}
\alias{WW_Watershed}
\title{Full watershed waterways}
\format{
a sf tibble with 439 rows and 28 columns.
}
\usage{
WW_Watershed
}
\description{
Shapefile of waterways in the San Francisco Estuary and its watershed
}
\section{Credit}{

This shapefile includes data from CH2M Hill, and Wetland and Water Resources, CDFG Ecosystem Restoration Program as a part of ERP Stage 2 conservation strategy, and Dan Gillenwater.
}

\keyword{datasets}
