% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R_DSIBM}
\alias{R_DSIBM}
\title{Subregions from the Delta Smelt Individual Based Model}
\format{
a sf tibble with 41 rows and 4 columns.
\describe{
\item{TNMID}{TNMID}
\item{METASOURCE}{METASOURCE}
\item{SOURCEDATA}{SOURCEDATA}
\item{SOURCEORIG}{SOURCEORIG}
\item{SOURCEFEAT}{SOURCEFEAT}
\item{LOADDATE}{LOADDATE}
\item{GNIS_ID}{GNIS_ID}
\item{SubRegion}{SubRegion.}
\item{AREAACRES}{AREAACRES.}
\item{AREASQKM}{AREASQKM}
\item{STATES}{STATES}
\item{HUC10}{HUC10}
\item{HUTYPE}{HUTYPE}
\item{HUMOD}{HUMOD}
\item{GLOBALID}{GLOBALID}
\item{SHAPE_Leng}{SHAPE_Leng}
\item{SHAPE_Area}{SHAPE_Area}
\item{geometry}{Shapefile polygon coordinates.}
}
}
\usage{
R_DSIBM
}
\description{
Subregions from the Delta Smelt Individual Based Model
}
\section{Credit}{

This shapefile was created by Flow West, Rose et al. 2013, and Peterson et al. 2019.
Peterson, J., E. McCreless, A. Duarte, S. Hamilton, and J. Medellin-Azuara. 2019. Structured Decision Making for Scientific Management in the San Francisco Bay-Delta. Draft progress report, US Bureau of Reclamation, Sacramento, CA. Data available on: https://github.com/CSAMP/delta-secchi-temperature-data.
Rose, K. A., W. J. Kimmerer, K. P. Edwards, and W. A. Bennett. 2013. Individual-Based Modeling of Delta Smelt Population Dynamics in the Upper San Francisco Estuary: I. Model Description and Baseline Results. Transactions of the American Fisheries Society 142:1238–1259.
}

\keyword{datasets}
