% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stations}
\alias{Stations}
\alias{WW_Watershed}
\title{Stations}
\format{
a sf tibble with 954 rows and 7 columns.
\describe{
\item{Station}{Code used for station.}
\item{StationName}{Full name of station (if applicable).}
\item{Latitude}{Latitude in decimal degrees, WGS84.}
\item{Longitude}{Longitude in decimal degrees, WGS84.}
\item{Parameter}{Type of data collected.}
\item{Source}{Name of sampling program.}
\item{geometry}{Shapefile point coordinates.}
}
Full watershed waterways

Shapefile of waterways in the San Francisco Estuary and its watershed

a sf tibble with 439 rows and 28 columns.
}
\usage{
WW_Watershed
}
\description{
Shapefile with regular sampling locations for most IEP surveys used between 2015 and 2021
}
\section{Credit}{

This shapefile includes data from CH2M Hill, and Wetland and Water Resources, CDFG Ecosystem Restoration Program as a part of ERP Stage 2 conservation strategy, and Dan Gillenwater.
}

\keyword{datasets}
